@@ -25,6 +25,8 @@
     "deploy": "gh-pages -d build -m '[ci skip] Updates'"
   },
   "dependencies": {
+    "@headlessui/react": "^2.1.2",
+    "@heroicons/react": "^2.1.5",
     "@polkadot/api": "^10.6.1",
     "@polkadot/extension-dapp": "^0.46.2",
     "@polkadot/keyring": "^12.1.2",
@@ -34,24 +36,29 @@
     "@polkadot/ui-settings": "^3.3.1",
     "@polkadot/util": "^12.1.2",
     "@polkadot/util-crypto": "^12.1.2",
+    "@radix-ui/themes": "^3.1.3",
     "node-polyfill-webpack-plugin": "^1.1.4",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-copy-to-clipboard": "^5.0.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.1",
     "react-scripts": "^5.0.0",
     "semantic-ui-css": "Semantic-Org/Semantic-UI-CSS#master",
     "semantic-ui-react": "^2.1.2",
     "stream-browserify": "^3.0.0"
   },
   "devDependencies": {
+    "autoprefixer": "^10.4.20",
     "eslint": "^8.10.0",
     "eslint-config-prettier": "^8.4.0",
     "eslint-config-react-app": "^7.0.0",
     "gh-pages": "^3.2.3",
+    "postcss": "^8.4.41",
     "prettier": "2.5.1",
     "react-app-rewired": "^2.2.1",
-    "semistandard": "^16.0.1"
+    "semistandard": "^16.0.1",
+    "tailwindcss": "^3.4.10"
   },
   "resolutions": {
     "eslint-plugin-jsx-a11y": "6.5.1"
