@@ -1,11 +1,13 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
-
+import ReactDom from 'react-dom/client';
 import App from './App'
+import { BrowserRouter } from "react-router-dom";
 
-ReactDOM.render(
+
+ReactDom.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <BrowserRouter>
+    <App/>
+    </BrowserRouter>
+  </React.StrictMode>
 )
